refactor: ViewSize 내부에 있던 프로토콜들을 별도의 파일로 분리 (ios-h/airbnb#38)
