+++ iOS/airbnb/Common/Protocol/ViewSizeApplicable.swift
@@ -1,0 +1,13 @@
+//
+//  ViewSize.swift
+//  airbnb
+//
+//  Created by 안상희 on 2022/06/06.
+//
+
+import UIKit
+
+protocol ViewSizeApplicable {
+    var width: CGFloat { get }
+    var height: CGFloat { get }
+}
